pyproject.toml
[tool.poetry]
name = "minos_auth"
version = "0.0.1"
description = "Minos Boilerplate contains all the boilerplate you need to create a Minos Python package."
readme = "README.md"
repository = "https://github.com/clariteia/minos_auth"
homepage = "https://clariteia.com"
authors = ["Clariteia Devs <devs@clariteia.com>"]
license = "MIT"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]
keywords = [
    "clariteia",
    "minos",
    "microservice",
    "auth",
]
packages = [
    { include = "minos" }
]
include = [
    "AUTHORS.md",
    "HISTORY.md",
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.3.2"
cached-property = "^1.5.2"
aiohttp-middlewares = "^1.1.0"
aiomisc = "^15.2.16"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^19.10b"
isort = "^5.8.0"
pytest = "^6.2.4"
coverage = "^5.5"
flake8 = "^3.9.2"
Sphinx = "^4.0.1"
pre-commit = "^2.12.1"
sphinx-autodoc-typehints = "^1.12.0"
sphinxcontrib-apidoc = "^0.3.0"
sphinx-rtd-theme = "^0.5.2"
m2r2 = "^0.2.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
auth = "minos.auth.cli:main"
